+++ lib/src/query_parser/free_text_query.dart
@@ -18,11 +18,18 @@ abstract class FreeTextQuery {
       required Iterable<QueryTerm> queryTerms,
       int targetResultSize = 50,
       WeightingStrategy weightingStrategy = WeightingStrategy.simple,
-      double? iDfThreshold}) {
-    return _FreeTextQueryImpl(
-        phrase, queryTerms.toSet(), targetResultSize,
-        weightingStrategy, iDfThreshold);
+      double? iDfThreshold,
+      bool expandUnmatched = false}) {
+    return _FreeTextQueryImpl(phrase, queryTerms.toSet(), targetResultSize,
+        weightingStrategy, iDfThreshold, expandUnmatched);
   }
+
+  /// The unmodified search phrase, including all modifiers and tokens.
+  String get phrase;
+
+  /// A flag that instructs search classes whether to expand any unmatched
+  /// terms (e.g. correct spelling).
+  bool get expand;
 
   /// A threshold value for the minimum inverse document frequency (iDf) of
   /// query terms.
@@ -30,9 +37,6 @@ abstract class FreeTextQuery {
   /// When [iDfThreshold] is not null and greater than 0.0, query terms that
   /// have a iDf less that [iDfThreshold] will be ignored.
   double? get iDfThreshold;
-
-  /// The unmodified search phrase, including all modifiers and tokens.
-  String get phrase;
 
   /// The desired size of the result set.
   int get targetResultSize;
@@ -54,32 +58,34 @@ abstract class FreeTextQuery {
 
   /// A hashmap of [QueryTermModifier] to weighting value.
   ///
-  /// Weights are used in calculating the [SearchResult.tfIdfScore], used in
+  /// Weights are used in calculating the [QuerySearchResult.tfIdfScore], used in
   /// extracting the top [FreeTextQuery.targetResultSize] results from an
   /// index.
   WeightingStrategy get weightingStrategy;
 
-  /// A hashmap of query terms to tf-idf for the term in the document.
+  /// A hashmap of query terms to weighted tf-idf for the term in the document.
+  /// The raw tf-idf is multiplied by the modifier weight and the phrase length
+  /// weight in the [weightingStrategy].
   ///
   /// - [dftMap] is a hashmap of query term to document frequency of the term
   ///   in the collection; and
   /// - [n] is the total number of documents in the collection.
   /// The tf-idf weighting for the document is the product of the term's
   /// frequency in the document with the inverse document frequency of
-  /// the term in the collection.
+  /// the term in the collection and the .
   ///
   /// ``` dart
   ///   tfIdf(idFt) => tFt * idFt
   /// ``
   Map<Term, double> tfIdfVector(DftMap dftMap, int n);
 
-  /// Returns a copy of the [FreeTextQuery], replacing elements of [queryTerms]
-  /// with values in [newTerms]. The key of the [newTerms] hashmap is the old
-  /// query term's term.
+  /// Replaces elements of [queryTerms] with the values in [newTerms].
+  ///
+  /// The key of the [newTerms] hashmap is the old query term's term.
   void expandTerms(Map<String, Iterable<QueryTerm>> newTerms);
 
-  /// Returns a copy of the [FreeTextQuery], removing elements of [queryTerms]
-  /// where the inverse document frequency is below [iDfThreshold]
+  /// Removes elements of [queryTerms] where the inverse document frequency is
+  /// below [iDfThreshold].
   void purgeTerms(DftMap docFrequencyMap, int n);
 }
 
@@ -121,11 +127,19 @@ abstract class FreeTextQueryMixin implements FreeTextQuery {
   @override
   Map<Term, double> tfIdfVector(DftMap dftMap, int n) {
     final termFrequencies = <String, int>{};
+    final qtMap = queryTermsMap;
     for (final t in allTerms) {
       termFrequencies[t] = (termFrequencies[t] ?? 0) + 1;
     }
-    return termFrequencies
-        .map((term, tF) => MapEntry(term, tF * dftMap.getIdFt(term, n)));
+    final Map<Term, double> retVal = {};
+    for (final qt in qtMap.values) {
+      final t = qt.term;
+      final wM = weightingStrategy.getWeight(qt).abs();
+      final tf = termFrequencies[t] = (termFrequencies[t] ?? 0);
+      final tfIdf = tf * dftMap.getIdFt(t, n);
+      retVal[t] = tfIdf * wM;
+    }
+    return retVal;
   }
 
   @override
@@ -159,8 +173,13 @@ class _FreeTextQueryImpl extends FreeTextQueryBase {
   ///   tokens; and
   /// - [queryTerms] is the a collection of all [QueryTerm] extracted from the
   ///   [phrase].
-  _FreeTextQueryImpl(this.phrase, Set<QueryTerm> queryTerms,
-      this.targetResultSize, this.weightingStrategy, this.iDfThreshold) {
+  _FreeTextQueryImpl(
+      this.phrase,
+      Set<QueryTerm> queryTerms,
+      this.targetResultSize,
+      this.weightingStrategy,
+      this.iDfThreshold,
+      this.expand) {
     this.queryTerms.addAll(queryTerms);
   }
 
@@ -169,4 +188,7 @@ class _FreeTextQueryImpl extends FreeTextQueryBase {
 
   @override
   final int targetResultSize;
+
+  @override
+  final bool expand;
 }